DD-2889 Refactor entities folder, correct manifests
