+++ bsc-subgraph.template.yaml
@@ -18,10 +18,6 @@ dataSources:
       abis:
         - name: ERC20Detailed
           file: ./abis/ERC20Detailed.json
-        - name: StakingPool
-          file: ./abis/StakingPool.json
-        - name: BasisVaultContract
-          file: ./abis/BasisVault.json
       eventHandlers:
         - event: Transfer(indexed address,indexed address,uint256)
           handler: handleTransfer
@@ -65,6 +61,8 @@ templates:
       abis:
         - name: BasisVault
           file: ./abis/BasisVault.json
+        - name: ERC20Detailed
+          file: ./abis/ERC20Detailed.json
       eventHandlers:
         - event: Deposit(indexed address,uint256,uint256)
           handler: handleDeposit
