+++ src/entities/globalStats/loadGlobalStats.ts
@@ -1,5 +1,3 @@
-import { BigInt } from "@graphprotocol/graph-ts";
-
 import { GlobalStat } from "../../../generated/schema";
 import { SINGLE_ENTITY_ID } from "../../utils";
 
@@ -10,7 +8,6 @@ export function loadGlobalStat(): GlobalStat {
     stats = new GlobalStat(SINGLE_ENTITY_ID);
 
     stats.activeMembersCount = 0;
-    stats.totalYieldEarned = BigInt.fromI32(0);
     stats.save();
   }
