+++ src/entities/shared/loadUser.ts
@@ -1,5 +1,5 @@
 import { Address } from "@graphprotocol/graph-ts";
-import { User } from "../../generated/schema";
+import { User } from "../../../generated/schema";
 
 export function loadOrCreateUser(address: Address): User {
   let id = address.toHex();
